refactor: make the code parameter of safeReject non-nullable (#2875)
The reject method of Promise class became non-nullable in React Native
0.76 after rewriting it from Java to Kotlin. We fixed it being called
with a nullable value in
https://github.com/hyochan/react-native-iap/pull/2872 by providing and
empty string as a fallback if it happens to be null. But after more
investigation it seems like it's not called with a value that could be
null anywhere in the code, so a better solution would be to just update
the type of the `code` parameter to be non-nullable.